feat: Bump dependencies and format code
This commit updates the following dependencies:
- sdk: ^3.7.0 -> ^3.8.0
- dart_style: ^3.0.1 -> ^3.1.0
- lints: ^5.1.1 -> ^6.0.0 (and in example)

It also formats the codebase using the updated `dart_style` and configures the formatter in `analysis_options.yaml` for both the main project and the example.
Specifically, `trailing_commas` is set to `preserve` and `page_width` to `120` (or `80` in `builder_impl.dart` via a comment).

Additionally, some minor code refactoring was done for brevity and clarity in `bin/lean_builder.dart`, `lib/src/element/element_impl.dart`, and `lib/src/resolvers/resolver.dart`.